+++ backend/canisters/registry/impl/src/lifecycle/init.rs
@@ -11,7 +11,7 @@ use utils::env::Environment;
 #[trace]
 fn init(args: Args) {
     canister_logger::init(args.test_mode);
-    init_cycles_dispenser_client(args.cycles_dispenser_canister_id);
+    init_cycles_dispenser_client(args.cycles_dispenser_canister_id, args.test_mode);
 
     let env = init_env([0; 32]);
     let mut data = Data::new(