+++ frontend/app/src/theme/types.ts
@@ -11,6 +11,9 @@ export interface Theme {
     burst: boolean;
     logo: boolean;
     mode: "light" | "dark"; // pick the one that makes the most sense for your theme
+    font: string;
+    "font-bold": string;
+    fontUrl?: string; // allows the dynamic loading of custom fonts
 
     bg: string;
     txt: string;
@@ -111,6 +114,10 @@ export interface Theme {
         bg: string;
 
         left: {
+            bg: string;
+        };
+
+        nav: {
             bg: string;
         };