+++ frontend/openchat-agent/src/services/community/candid/types.d.ts
@@ -581,6 +581,7 @@ export interface DiamondMembershipDetails {
   'expires_at' : TimestampMillis,
 }
 export type DiamondMembershipPlanDuration = { 'OneYear' : null } |
+  { 'Lifetime' : null } |
   { 'ThreeMonths' : null } |
   { 'OneMonth' : null };
 export type DirectChatCreated = {};